Fixed bugs in changing avatars
Restored default-project's onUpdate eventHook. It now adds new avatars
it contains, and removes any existing avatars that were in it that no
longer have any associated static-resources.

Updated avatar.hooks.ts to use raw _patch on userAvatarPath for avatar
deletion, as prior implementation did not work for internal calls that
do not have a user associated with them.

Resolves IR-7214
